Cleaning up code not yet ready
